Move formatting to stats module
