+++ index.js
@@ -12,6 +12,7 @@ const format = require('./lib/format')
 const createConfController = require('./controllers/createConfController')
 const landingController = require('./controllers/landingController')
 const sendValidationEmailController = require('./controllers/sendValidationEmailController')
+const stats = require('./lib/stats')
 const urls = require('./urls')
 
 const app = express()
@@ -77,16 +78,18 @@ app.get(urls.legalNotice, (req, res) => {
 
 app.get(urls.stats, async (req, res) => {
   const NUM_STATS_POINTS = 20
-  let stats = []
+  let latestStats = []
   try {
-    stats = await db.getLatestStatsPoints(NUM_STATS_POINTS)
+    latestStats = await db.getLatestStatsPoints(NUM_STATS_POINTS)
   } catch (err) {
     console.error(`Impossible de récupérer les statsPoints`, err)
   }
 
+  const formattedStats = stats.formatDataForDisplay(latestStats)
+
   res.render('stats', {
     pageTitle: 'Statistiques',
-    stats: stats,
+    stats: formattedStats,
   })
 })
