+++ lib/stats.js
@@ -38,3 +38,20 @@ module.exports.computeStats = async () => {
   const statsPoint = await getOnlineParticipantStats()
   db.insertStatsPoint(statsPoint)
 }
+
+/**
+ * inData coming from db is like :
+ * [{"date":"2020-11-05T14:11:00.440Z","onlineParticipantsCount":0,"activeConfsCount":0,"errorConfsCount":0}, {...}, ...]
+ * outData is formatted like :
+ * { labels: [dates], onlineParticipantsSeries: [series], activeConfsSeries: [series] }
+ */
+module.exports.formatDataForDisplay = (inData) => {
+  const outData = { labels: [], onlineParticipantsSeries: [], activeConfsSeries: [] }
+  inData.forEach(dataPoint => {
+    // Use unshift to add at the beginning of array, because the inData is in reverse chronological order.
+    outData.labels.unshift(new Date(dataPoint.date))
+    outData.onlineParticipantsSeries.unshift(dataPoint.onlineParticipantsCount)
+    outData.activeConfsSeries.unshift(dataPoint.activeConfsCount)
+  })
+  return outData
+}
